AUTHORS.rst: Document creators
I think it's important there are no assumptions when reading the
document. There should be nothing wrong to celebrate the people who
created the project, but we also need to keep in mind how we can
welcome new people by ensuring that they have a path to become
equally influential.