Backward incompatible: *.send now requires keyword-only arguments only
I keep making the mistake of doing:

    channel.send(x)

and expect that to send `x` as the value.

But the signature is `(key, value, ...)`, so it ends up being
`channel.send(key=x, value=None)`.

This is confusing to other users as well, most recently in post
on the faust-users mailing-list.

Fixing this will come in two parts:

1) Faust 1.1 (this change): Make them keyword-only arguments

    This will make it an error if the names of arguments are not
    specified:

        channel.send(key, value)

    Needs to be changed to:

        channel.send(key=key, value=value)

2) Faust 1.2: We will change the signature to ``channel.send(value, key=key, ...)``

    At this stage all existing code will have changed to using
    keyword-only arguments.