Implements support for model schemas (Issue #315)
This implements a single structure (Schema) that configures
the key_type/value_type/key_serializer/value_serializer for a topic or
agent.

The benefit of having an abstraction a level above codecs
is that schemas can implement support for serialization formats
such as ProtocolBuffers, Apache Thrift and Avro.

The schema will also have access to the Kafka message headers,
necessary in some cases where serialization schema is specified
in headers.