Models: Support arbitrarily nested type expressions.
This extends model fields to support arbitrarily nested type
expressions, such as `List[Dict[str, List[Set[MyModel]]]`

In addition it also adds support for:

- Named tuples in expressions (including namedtuples where fields are
  models).
- Unions of models `Union[ModelX, ModelY]`

Important Notes
===============

One important change in this diff is that serializing a model will
no longer traverse the structure for child models, instead we rely
on the json serializer to call `Model.__json__()` during serializing.

Specifically this means, where previously having models

```py

class X(Model):
    name: str

class Y(Model):
    x: X

and calling `Y(X('foo')).to_representation()` it would return:

```py

>>> Y(X('foo')).to_representation()
{
    'x': {
        'name': 'foo',
        '__faust': {'ns': 'myapp.X'},
    },
    '__faust': {'ns': 'myapp.Y'},
}
```

after this change it will instead return the objects as-is:

```py

>>> Y(X('foo')).to_representation()
{
    'x': X(name='foo'),
    '__faust': {'ns': 'myapp.Y'},
}
```

This is a backward incompatible change for anything that relies
on the previous behavior, but in most apps will be fine as the
Faust json serializer will automatically handle models and call
`Model.__json__()` on them as needed.

This will also speed up model serialization slightly, so it's an
improvement we have been planning for a while.

Removed attributes
------------------

The following attributes have been removed from `Model._options`,
and `FieldDescriptorT`, as they are no longer needed,
or no longer make sense when supporting arbitrarily nested structures.

`Model._options`
~~~~~~~~~~~~~~~~

- `models`

    Previously map of fields that have related models.
    This index is no longer used, and a field can have multiple
    related models now.  You can generate this index using the
    statement:

    ```py

        {field: field.related_models
            for field in model._options.descriptors
            if field.related_models}

    ```

- `modelattrs`

- ``field_coerce``

- ``initfield``

- ``polyindex``

`FieldDescriptorT`

- `generic_type`
- `member_type`

Implementation Details
----------------------

When creating a model we analyze each field and
compile an expression that can be used to deserialize the field from
serialized data.

For example for the type expression ``List[List[MyModel]]`` we
use code generation to build a function that deserialize the data
using a comprehension:

```py
>>> from faust.models.typing import TypeExpression

>>> expr = TypeExpression(List[List[MyModel]])
>>> fun = expr.as_function()
>>> print(fun.__sourcecode__)
def expr(a):
  return [[myapp__MyModel._from_data_field(c) for c in b] for b in a]
```

This generated function is stored on the field descriptor
as ``FieldDescriptor._to_python()`` and ``Model.from_data()`` will
call this for every field.