Consumer: yield back to the event loop for every 500 messages consumed
This fixes a problem of the worker blocking the event loop for too long.

A common misconception about how asyncio works is that
many think the ``await`` statement will yield back to the event loop
each time it is called. but this is not how it works.

In asyncio any running ``async def`` function will be converted
into a task, so if we have the following chain of functions:

```
async def foo():
   return await bar()

async def bar():
    return await baz()

async def baz():
    await asyncio.sleep(0)
    await xuzzy()

async def xuzzy():
    return 303

loop = asyncio.get_event_loop()
loop.run_until_complete(foo())
```

The event loop adds ``foo()`` as a pending task and will periodically
advance the underlying generator of foo():

```
class Task:

    def __init__(self, coro: Coroutine) -> None:
        self.coro = coro
        self.loop = asyncio.get_event_loop()

    def __step(self):
        try:
            result = self.coro.send(None)  # advance the generator
        except StopIteration:
            self.loop.call_soon(self.set_result, result)
        except BaseException as exc:
            # mark the task as failed with exception
            self.loop.call_soon(self.set_exception, exc)
        else:
            # still more to do so schedule our coruotine
            # to be advanced on the next iteration of the loop.
            self.loop.call_soon(self.__step)
```

So where will the coroutine ``foo()`` abve yield back to the event loop?
That is, how much of foo() will excecute in a single __step in the event
loop.  It will actually execute all of it until it hits the
``asyncio.sleep()`` (or any I/O that will block such as a network
request).

This means that a chain of couroutines being awaited can easily
block the event loop for a long time, if it does not do any real network
I/O.

We have to monitor for any tasks where a single step
takes too long, and intervene by adding manual ``asyncio.sleep(0)``
calls.